Disable check for C++ compiler
By default, CMake assumes that a project uses both the C and C++
languages, and unionfs-fuse's CMakeLists.txt doesn't explicit which
languages it uses. Therefore CMake checks for both a C compiler and a
C++ compiler, which prevents from building unionfs-fuse on platforms
that don't have a C++ compiler.

We adjust unionfs-fuse CMakeLists.txt to explicity state that we only
use the C language, so that the C++ compiler check is no longer
performed by Cmake.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>